+++ changelog.md
@@ -1,3 +1,5 @@
+- add ability to add arbitrary attributes to images & links [#5]
+
 # 0.6.0
 - allow looser set of contents for merge tags
 - change to passing options to HeaderWithID, now a `header_with_id:` key and options for `:slug` and `:level`